Offload commit encoding entirely to git
Commits can be encoding with different character sets, which means that
relying on a repository wide setting is wrong. To fix this pass
--encoding=UTF-8 to all git commands, which use the revision machinery.
This will override the user's i18n.logOutputEncoding option and ensure
that the output provided by git is encoding as UTF-8. The --encoding
option has been available in git for a couple of years--since 2006 for
the log and diff related tools and since 2008 for git-blame--so they
should not pose any problems.

This patch is based on work by Heiko Schlichting.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>