prompt_menu: protect against programming error
If prompt_menu is passed an empty menu list, you can get negative
values for selected in the KEY_{LEFT,UP} case, and a modulo by
zero in the KEY_{RIGHT,DOWN} cases.

This currently can't happen, because the only user of prompt_menu is
toggle_option that does fill up the menu. However, let's future-proof
the call: since an empty menu would indicate a programming error, assert
that size must be positive.

Signed-off-by: Giuseppe Bilotta <giuseppe.bilotta@gmail.com>