[GH #70] Fix parent blaming when tig is launched in subdirectory
When running "tig blame" from within a subdirectory of a repository, the
"B" and "," commands will fail with "No blame exist for...". The problem
is that the current directory name gets prepended to the filename if
view->prev == NULL, but it should not be in this case.

Steps to reproduce:
1) checkout a copy of tig
2) cd contrib
3) run "tig blame tig-completion.bash"
4) press B
5) observe "tig: No blame exist for contrib/contrib/tig-completion.bash"

Also fixes another minor bug in an expression.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>