fix: don't add one to make display consistent
The option value specifies the length including the padding character. I
initially added one to mean: "if I specify an id-len of 7, I get 7 hex chars".
However it seems more sensible to make this option consistent with the others
to mean: "If i specify an id-len of 7, I get 6 hex chars plus a single
whitespace for padding."