read core.abbrev, but prefer id-width
There are now two settings that can affect how long the commit ID in the main
view is: 'core.abbrev' and 'id-width'. The first is a core option, the second
is a tig specific one. In case tig finds the first setting it respects this. In
case the user wishes tig to use a different value, the second option is
provided as an oevrride.