fix: reading git colors in rang 0 255
In 'tig' a color from the 256 colors range is specified using the string
'colorN' where N is the integer. In 'git' you can used 256 colors too. However
instead of 'colorN' they are specified using just N. This needs to be taken
into account for using the option 'use-git-colors' which reads options from the
git config for use in tig.