Don't clear the status/report line when finishing a piped update
This appears to be no longer necessary, and a hold over from when
updates were more verbose at the start of and during processing and
needed to be cleared when finished (refer to f97f401).

More importantly, the call to report_clear() when the view is done
updating can hide status messages from the beginning of the update which
should remain visible past the end (until some other action clears the
status line).

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>