main: add options for short commit titles
Many developers and tools use the Git convention for commit titles,
as discussed in the git-commit(1) man page:

  Though not required, it's a good idea to begin the commit message with
  a single short (less than 50 character) line summarizing the change

This patch adds the ability to highlight text exceeding a specified
width in commit titles, through a "show-short-title" toggle.

It also adds a "short-title-width" option to change the maximum value to
consider a commit title to be short. This value defaults to 50.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>