main: add toggle for commit title overflow
Many developers and tools use the Git convention for commit titles,
as discussed in the git-commit(1) man page:

  Though not required, it's a good idea to begin the commit message with
  a single short (less than 50 character) line summarizing the change

This patch adds the ability to highlight text exceeding a specified
width in commit titles, through a "title-overflow" option.

As a boolean, it enables/disables the highlighting, with a default value
of 50. As a int, it enables and uses the value as the maximum width for
a title to be considered short.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>