Provide mkstemps() work-around for platforms lacking it
The function mkstemps() isn't available in all libc implementations. In
glibc it first became available in 2.11. This is likely true of many
non-LINUX platforms as well.

The work-around is based on mkstemps() implementation from libiberty and
is kept inside its own source file under a new compat/ subdirectory. The
work-around can be enabled via configure (and config.make.in), via the
config.make file by adding the line 'NO_MKSTEMPS=y' or when calling make
by adding 'NO_MKSTEMPS=y' to its parameter list.

This fixes breakage that was introduced with a0fdac29 "Create temporary
file with name as suffix."

Fixes #152
Closes #175

Signed-off-by: Drew Northup <n1xim.email@gmail.com>

Original patch by Drew reworked to make it easier to enable the
work-around when not using configure. Also, switched mkstemps()
implementation to that of libiberty instead of the adapted version from
Git.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>