Refactor the pager mode handling
Remove the global opt_stdin variable and the VIEW_STDIN flag and move
checking of pager mode options and opening logic into open_pager_mode().
Two new flags (OPEN_PAGER_MODE and OPEN_FORWARD_STDIN) is added to deal
with how the IO for the view needs to be initialized.

Also fixes a bug, where stdin was not closed when pager mode is not
supported. This means that the process writing to stdin would hang until
Tig closes.