Add "auto" vertical-split
This patch adds a new option to control the vertical split.
If the value of the "vertical-split" option is set to "auto" (which is
the default behavior as of this commit), Tig will choose either vertical
or horizontal split depending on the window size.

Forcing the "vertical-split" option to "yes" or "no" naturally disables
the auto feature.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>
Signed-off-by: Jonas Fonseca <fonseca@diku.dk>