graph: uncomment is_boundary propagation
This code came into being commented out in 607cfb1 (Rewrite
the revision graph renderer, 2009-12-11), but it is not
clear why (perhaps because of the s/->/./ compile error, and
was meant to be looked at later?).

As a result, we would lose the is_boundary flag, leading to
them being displayed as normal commits.  Probably nobody
noticed, though, because boundary commits were otherwise
broken until the previous commit.