prompt: introduce :toggle command
This moves most toggle actions to use :toggle <option-name>. This new
:toggle prompt command enables to manipulate all the current options
while avoiding having to create a separate toggle-* action for each
option.  To achieve this the :toggle command changes behavior based on
the type of the option that is toggled:

 - For bool and enum options it works the similar to how toggle-*
   actions works today.
 - For argument array options (e.g. diff-options), it permits to
   append/remove one or more arguments (similar to how
   toggle-ignore-space works).
 - For int and double options it allows to increment and decrement the
   value (e.g. author-width).

Bindings has been added to tigrc to stay backwards compatible.

Fixes #69