Allow prompt toggle commands to both reset display and reload
When changing the vertical split setting the diff stat was not
automatically updated to fit the full length of the view. By introducing
a proper VIEW_RESET_DISPLAY flag and using that when toggling the
vertical-split option the diff view is now automatically reloaded so
that the diff stat has the correct width.