Fix readline support
The watch code introduced a bug where stdin was marked with nodelay()
when reading multibyte key input. This changed get_input() to use
wtimeout() instead of nodelay(). Unfortunately, as a defensive measure
an extra nodelay() was added before the loop that reads the remaining
multibyte stream, which in turn broke readline support.