Make blame ID colors predictable across reloads
Change draw_id() to use a hash function to generate a color index based
on the ID characters so that the same commit ID is always assigned the
same color across reloads. Previously the memory location of the commit
struct's ID member was used to derive the color index.

To further improve the ID coloring, increase the palette to 14 colors by
adding "boldified" versions of the base ASCII colors.