Add workaround for git pretty.format settings that break the log view
This overrides any pretty.format setting that contains color definitions
to use the 'medium' format. The fix adds the setting 'log-options' for
implementing the workaround. The option can be configured in `tigrc` and
works similar to `blame-options`. Finally, in order to test the feature
the semantics of `:toggle <args-option>` (note no arguments) has been
changed to simply clear the entries in the <args-option>.

Fixes #225