Fix status view regression where tracked files was shown as added
The changes in 3615bdee4c3d1a0c719fc6b24f25f5745a6f2d92 introduced the
regression. Since it only occurs when `tig status` is run in a repo with
a detached head it was not caught by status/on-branch-test. To rectify
this on-branch-test now checks each individual test case as well as
running them all together.