Workaround segfault when refreshing the status view
When refreshing the status view, it will try to readd the HEAD
reference. This caused a corruption in the reference DB, which led to a
segfault when returning to the main view.

The workaround will simply skip any updates to the ID reference list
when the ID does not change for an existing ref. More work is needed to
properly update the ID reference list when changes occur.

References #370.