Use commit-order setting for controlling automatic topo-ordering
Remove the `commit-title-graph=no-topo` option and use `commit-order`
for setting whether or not to automatically use topo-ordering when the
graph is enabled. The main reason is that `commit-title-graph` is meant
for setting whether the graph is shown, whereas `commit-order` is the
perfect fit.

By default, `commit-order` is now set to "auto" which will do this.
However, users can set it to "default" (or any other commit order) to
disable this behavior. This matches the behavior of similar enum
settings that also have an "auto" value.

With this the main issues related with the performance degredation in
large repositories have been addressed and the two

Fixes #310
Fixes #324